+++ bench/Cargo.toml
@@ -12,7 +12,7 @@ name = "main"
 path = "./src/main.rs"
 
 [dependencies]
-p2p = { path = "..", package = "p2p" }
+p2p = { path = "..", package = "tentacle" }
 rand = "0.6.1"
 tokio = "0.1.14"
 futures = "0.1.25"
