+++ ping/Cargo.toml
@@ -6,7 +6,7 @@ authors = ["Nervos Core Dev <dev@nervos.org>"]
 edition = "2018"
 
 [dependencies]
-p2p = { path = "..", package = "p2p" }
+p2p = { path = "..", package = "tentacle" }
 tokio = "0.1"
 log = "0.4"
 flatbuffers = "0.5.0"
